Move find_with_contours and adjust import
