+++ OpticalNavigation/core/find_algos/find_with_contours.py
